Specify what folder to upload the track to - Music or Video.
2008-04-16  Gabriel Burt  <gabriel.burt@gmail.com>

	* src/Dap/Banshee.Dap.Mtp/Banshee.Dap.Mtp/MtpSource.cs: Specify what
	folder to upload the track to - Music or Video.

	* src/Libraries/Mtp/Mtp/MtpDevice.cs: Set the parent_id on the track
	before uploading it.

	* src/Libraries/Mtp/Mtp/Track.cs: Improve the DetectFileType method.

	* src/Libraries/Hyena/Hyena.Query/FileSizeQueryValue.cs:
	* src/Libraries/Hyena/Hyena.Query/RelativeTimeSpanQueryValue.cs:
	* src/Libraries/Hyena/Hyena.Query/TimeSpanQueryValue.cs: Parse values as
	doubles instead of ints, allowing queryies and smart playlists to specy
	"size>2.2MB" for instance.

	* src/Libraries/Hyena.Gui/Hyena.Query.Gui/FileSizeQueryValueEntry.cs:
	* src/Libraries/Hyena.Gui/Hyena.Query.Gui/TimeSpanQueryValueEntry.cs: Add
	a single decimal place to the spin button.

	* src/Libraries/Hyena/Hyena/StringUtil.cs: Improve FormatDouble method,
	now returns 1 if N < 1.1, etc.

svn path=/trunk/banshee/; revision=3785

+++ src/Libraries/Hyena/Hyena.Query/RelativeTimeSpanQueryValue.cs
@@ -55,11 +55,11 @@ namespace Hyena.Query
             get { return "date"; }
         }
 
-        public override long Offset {
+        public override double Offset {
             get { return -offset; }
         }
 
-        public override void SetUserRelativeValue (long offset, TimeFactor factor)
+        public override void SetUserRelativeValue (double offset, TimeFactor factor)
         {
             SetRelativeValue (-offset, factor);
         }
@@ -75,7 +75,7 @@ namespace Hyena.Query
             return DateTimeUtil.FromDateTime (DateTime.Now + TimeSpan.FromSeconds ((double) offset)).ToString (System.Globalization.CultureInfo.InvariantCulture);
         }
 
-        protected override string FactorString (TimeFactor factor, int count)
+        protected override string FactorString (TimeFactor factor, double count)
         {
             string translated = base.FactorString (factor, count);
             return (translated == null) ? null : String.Format (Catalog.GetString ("{0} ago"), translated);
